restructure join class request view
