+++ src/pages/teacherDashboard/classes/Classes.tsx
@@ -5,7 +5,7 @@ import { type SchoolTeacherUser } from "codeforlife/api"
 import Class from "./class/Class"
 import ClassTable from "./ClassTable"
 import CreateClassForm from "./CreateClassForm"
-import JoinClassRequest from "./JoinClassRequest"
+import JoinClassRequest from "./joinClassRequest/JoinClassRequest"
 import JoinClassRequestTable from "./JoinClassRequestTable"
 import ResetStudentsPassword from "./class/ResetStudentsPassword"
 import { type RetrieveUserResult } from "../../../api/user"
